organize deprecated serializers
